Separate libs and types
